fix flow issues and protocols test
