+++ src/protocols/login.test.js
@@ -23,6 +23,7 @@ const usrAddr = '0x308eff1357e7b5881c00ae22463b0f69a0d58adb';
 const proofEthName = {
   ethName: 'usertest@iden3.io',
   proofAssignName: {
+    signer: '0x0123456789abcdef0123456789abcdef01234567',
     date: 1551457426,
     leaf: '0x0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000308eff1357e7b5881c00ae22463b0f69a0d58adb0006e87e2d5265f1ea2ce3882864eb871263080c73ba1d4bfe89288a028f90300000000000000000000000000000000000000000000000000000000000000003',
     proofs: [
@@ -38,6 +39,7 @@ const proofEthName = {
 };
 
 const proofKSign = {
+  signer: '0x0123456789abcdef0123456789abcdef01234567',
   date: 1551457426,
   leaf: '0x000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003c2e48632c87932663beff7a1f6deb692cc61b041262ae8f310203d0f5ff50000000000000000000000000000000000007833000000000000000000000004',
   proofs: [
